backend/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Caching==2.1.0
Flask-RESTX==1.3.0
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.1
gunicorn==21.2.0

# Korean Market
FinanceDataReader==0.9.50
pykrx==1.0.46

# US Market
yfinance==0.2.37
lxml>=4.9.0
requests>=2.31.0

# Data Analysis
pandas==2.2.0
numpy==1.26.4

# Testing
pytest==8.0.0
pytest-flask==1.3.0
